# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/Treatments": {
            "get": {
                "description": "list Treatment entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List Treatment entities",
                "operationId": "list-Treatment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Treatment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Treatment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Treatment",
                "operationId": "create-Treatment",
                "parameters": [
                    {
                        "description": "Treatment entity",
                        "name": "Treatment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Treatment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Treatment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/Treatments/{id}": {
            "get": {
                "description": "get Treatment by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a Treatment entity by ID",
                "operationId": "get-Treatment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Treatment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Treatment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get Treatment by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a Treatment entity by ID",
                "operationId": "delete-Treatment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Treatment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.Treatment": {
            "type": "object",
            "properties": {
                "datetreat": {
                    "type": "string"
                },
                "doctor": {
                    "type": "integer"
                },
                "patientrecord": {
                    "type": "integer"
                },
                "treatment": {
                    "type": "string"
                },
                "typetreatment": {
                    "type": "integer"
                }
            }
        },
...
        "ent.Bill": {
            "type": "object",
            "properties": {
                "Amount": {
                    "description": "Amount holds the value of the \"Amount\" field.",
                    "type": "string"
                },
                "Date": {
                    "description": "Date holds the value of the \"Date\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BillQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BillEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Doctor": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DoctorQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DoctorEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DoctorEdges": {
            "type": "object",
            "properties": {
                "EdgesOfDoctorinfo": {
                    "description": "EdgesOfDoctorinfo holds the value of the EdgesOfDoctorinfo edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctorinfo"
                },
                "EdgesOfTreatment": {
                    "description": "EdgesOfTreatment holds the value of the EdgesOfTreatment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Treatment"
                    }
                },
                "EdgesOfUser": {
                    "description": "EdgesOfUser holds the value of the EdgesOfUser edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
        "ent.Doctorinfo": {
            "type": "object",
            "properties": {
                "doctorname": {
                    "description": "Doctorname holds the value of the \"doctorname\" field.",
                    "type": "string"
                },
                "doctorsurname": {
                    "description": "Doctorsurname holds the value of the \"doctorsurname\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DoctorinfoQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DoctorinfoEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "licensenumber": {
                    "description": "Licensenumber holds the value of the \"licensenumber\" field.",
                    "type": "string"
                },
                "telephonenumber": {
                    "description": "Telephonenumber holds the value of the \"telephonenumber\" field.",
                    "type": "string"
                }
            }
        },
        "ent.DoctorinfoEdges": {
            "type": "object",
            "properties": {
                "EdgesOfDepartment": {
                    "description": "EdgesOfDepartment holds the value of the EdgesOfDepartment edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Department"
                },
                "EdgesOfDoctor": {
                    "description": "EdgesOfDoctor holds the value of the EdgesOfDoctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                },
                "EdgesOfEducationlevel": {
                    "description": "EdgesOfEducationlevel holds the value of the EdgesOfEducationlevel edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Educationlevel"
                },
                "EdgesOfOfficeroom": {
                    "description": "EdgesOfOfficeroom holds the value of the EdgesOfOfficeroom edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Officeroom"
                },
                "EdgesOfPrename": {
                    "description": "EdgesOfPrename holds the value of the EdgesOfPrename edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Prename"
                }
            }
        },
...
        "ent.Gender": {
            "type": "object",
            "properties": {
                "Genderstatus": {
                    "description": "Genderstatus holds the value of the \"Genderstatus\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the GenderQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.GenderEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.GenderEdges": {
            "type": "object",
            "properties": {
                "EdgesOfPatientrecord": {
                    "description": "EdgesOfPatientrecord holds the value of the EdgesOfPatientrecord edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Patientrecord"
                    }
                }
            }
        },
        "ent.Historytaking": {
            "type": "object",
            "properties": {
                "bp": {
                    "description": "Bp holds the value of the \"bp\" field.",
                    "type": "integer"
                },
                "datetime": {
                    "description": "Datetime holds the value of the \"datetime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the HistorytakingQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.HistorytakingEdges"
                },
                "hight": {
                    "description": "Hight holds the value of the \"hight\" field.",
                    "type": "number"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "oxygen": {
                    "description": "Oxygen holds the value of the \"oxygen\" field.",
                    "type": "string"
                },
                "pulse": {
                    "description": "Pulse holds the value of the \"pulse\" field.",
                    "type": "integer"
                },
                "respiration": {
                    "description": "Respiration holds the value of the \"respiration\" field.",
                    "type": "integer"
                },
                "symptom": {
                    "description": "Symptom holds the value of the \"symptom\" field.",
                    "type": "string"
                },
                "temp": {
                    "description": "Temp holds the value of the \"temp\" field.",
                    "type": "number"
                },
                "weight": {
                    "description": "Weight holds the value of the \"weight\" field.",
                    "type": "number"
                }
            }
        },
...
        "ent.Medicalrecordstaff": {
            "type": "object",
            "properties": {
                "Name": {
                    "description": "Name holds the value of the \"Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the MedicalrecordstaffQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MedicalrecordstaffEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Patientrecord": {
            "type": "object",
            "properties": {
                "Age": {
                    "description": "Age holds the value of the \"Age\" field.",
                    "type": "integer"
                },
                "Allergic": {
                    "description": "Allergic holds the value of the \"Allergic\" field.",
                    "type": "string"
                },
                "Bloodtype": {
                    "description": "Bloodtype holds the value of the \"Bloodtype\" field.",
                    "type": "string"
                },
                "Date": {
                    "description": "Date holds the value of the \"Date\" field.",
                    "type": "string"
                },
                "Disease": {
                    "description": "Disease holds the value of the \"Disease\" field.",
                    "type": "string"
                },
                "Email": {
                    "description": "Email holds the value of the \"Email\" field.",
                    "type": "string"
                },
                "Home": {
                    "description": "Home holds the value of the \"Home\" field.",
                    "type": "string"
                },
                "Idcardnumber": {
                    "description": "Idcardnumber holds the value of the \"Idcardnumber\" field.",
                    "type": "integer"
                },
                "Name": {
                    "description": "Name holds the value of the \"Name\" field.",
                    "type": "string"
                },
                "Phonenumber": {
                    "description": "Phonenumber holds the value of the \"Phonenumber\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientrecordQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientrecordEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.PatientrecordEdges": {
            "type": "object",
            "properties": {
                "EdgesOfGender": {
                    "description": "EdgesOfGender holds the value of the EdgesOfGender edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Gender"
                },
                "EdgesOfHistorytaking": {
                    "description": "EdgesOfHistorytaking holds the value of the EdgesOfHistorytaking edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Historytaking"
                    }
                },
                "EdgesOfMedicalrecordstaff": {
                    "description": "EdgesOfMedicalrecordstaff holds the value of the EdgesOfMedicalrecordstaff edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Medicalrecordstaff"
                },
                "EdgesOfPatientrecordPatientrights": {
                    "description": "EdgesOfPatientrecordPatientrights holds the value of the EdgesOfPatientrecordPatientrights edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Patientrights"
                    }
                },
                "EdgesOfPrename": {
                    "description": "EdgesOfPrename holds the value of the EdgesOfPrename edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Prename"
                },
                "EdgesOfTreatment": {
                    "description": "EdgesOfTreatment holds the value of the EdgesOfTreatment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Treatment"
                    }
                }
            }
        },
        "ent.Patientrights": {
            "type": "object",
            "properties": {
                "PermissionDate": {
                    "description": "PermissionDate holds the value of the \"PermissionDate\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientrightsQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientrightsEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "insurance_id": {
                    "type": "integer"
                },
                "patientrightstype_id": {
                    "type": "integer"
                }
            }
        },
...
        "ent.Prename": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PrenameQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PrenameEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "prefix": {
                    "description": "Prefix holds the value of the \"prefix\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Treatment": {
            "type": "object",
            "properties": {
                "Datetreat": {
                    "description": "Datetreat holds the value of the \"Datetreat\" field.",
                    "type": "string"
                },
                "Treatment": {
                    "description": "Treatment holds the value of the \"Treatment\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the TreatmentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.TreatmentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.TreatmentEdges": {
            "type": "object",
            "properties": {
                "EdgesOfDoctor": {
                    "description": "EdgesOfDoctor holds the value of the EdgesOfDoctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                },
                "EdgesOfPatientrecord": {
                    "description": "EdgesOfPatientrecord holds the value of the EdgesOfPatientrecord edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patientrecord"
                },
                "EdgesOfTypetreatment": {
                    "description": "EdgesOfTypetreatment holds the value of the EdgesOfTypetreatment edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Typetreatment"
                },
                "EdgesOfUnpaybills": {
                    "description": "EdgesOfUnpaybills holds the value of the EdgesOfUnpaybills edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Unpaybill"
                }
            }
        },
        "ent.Typetreatment": {
            "type": "object",
            "properties": {
                "Typetreatment": {
                    "description": "Typetreatment holds the value of the \"Typetreatment\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the TypetreatmentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.TypetreatmentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.TypetreatmentEdges": {
            "type": "object",
            "properties": {
                "EdgesOfTreatment": {
                    "description": "EdgesOfTreatment holds the value of the EdgesOfTreatment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Treatment"
                    }
                }
            }
        },
        "ent.Unpaybill": {
            "type": "object",
            "properties": {
                "Status": {
                    "description": "Status holds the value of the \"Status\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UnpaybillQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UnpaybillEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.UnpaybillEdges": {
            "type": "object",
            "properties": {
                "EdgesOfBills": {
                    "description": "EdgesOfBills holds the value of the EdgesOfBills edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Bill"
                },
                "EdgesOfTreatment": {
                    "description": "EdgesOfTreatment holds the value of the EdgesOfTreatment edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Treatment"
                }
            }
        },
        "ent.User": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UserEdges"
                },
                "email": {
                    "description": "Email holds the value of the \"email\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "images": {
                    "description": "Images holds the value of the \"images\" field.",
                    "type": "string"
                },
                "password": {
                    "description": "Password holds the value of the \"password\" field.",
                    "type": "string"
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}